add assert for lb ub (#39)
* add assert for lb ub

Currently if lb and ub are switched, this will still silently return the wrong answer. Adding an assert here will catch this and warn the user of the error.

* Change assert to error throwing

Previous commit used an assert to warn the user about lb being larger than ub. This commit follows best practices and replaces the assert call with an error handling one.

* add check_bounds to avoid precomp + allocation

Added a check bounds function in order to have it precompile, and changed how bounds were checked in order to prevent allocations.

* run formatter

* Revert "run formatter"

This reverts commit f2fcc9522e5882922791b4402cf4dcd870454246.

* run formatter again

* run formatter...again

* fix check_bounds

* add dispatch for check_bounds

* fix check_bounds

* revert changes

* add inequality to fix check

* add tests

* Update test/runtests.jl

Co-authored-by: Christopher Rackauckas <accounts@chrisrackauckas.com>

* run formatter

Co-authored-by: Christopher Rackauckas <accounts@chrisrackauckas.com>